requirements.txt
timm
safetensors
pandas
numpy
scikit-learn
pyyaml
psutil
pillow
torch
torchvision